+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the Audio Player project will be documented in this file.
 ## 3.4.2 - 2025-07-12
 ### Fixed
 - PHP 8.4 compatibility for nullable parameters
+### Changed
+- Refactor SidebarController into service and mapper
 
 ## 3.4.1 - 2023-12-11
 ### Fixed